update Label & TaskStatus Controllers
- Конструктор контроллера вызывается в момент создания его экземпляра.
Laravel создает экземпляр контроллера, когда обрабатывает HTTP-запрос
(например отправленный пользователем через браузер) - проверяет маршруты
в файле маршрутов routes/web.php и определяет, какой контроллер и метод
вызвать для обработки.

После того как маршрутизация определила нужный контроллер и метод,
Laravel создает экземпляр этого контроллера.
После выполнения конструктора управление передается к методу, который
был вызван в маршруте.

Метод $this->authorizeResource(Label::class) устанавливает авторизацию
для всех действий в контроллере, используя Policy для проверки прав
доступа пользователя к ресурсам Label.

- В экшене destroy() проверка, существуют ли связанные задачи для данной
Label в базе и если сущ возвращает пользователя на пред страницу
(с которой был вызван метод) и добавляет ошибку в сессию, чтобы
отобразить её на странице.

+++ app/Http/Controllers/LabelController.php
@@ -12,7 +12,7 @@ class LabelController extends Controller
 {
     public function __construct()
     {
-        $this->middleware('auth')->except('index');
+        $this->authorizeResource(Label::class);
     }
 
     public function index(): View
@@ -48,6 +48,13 @@ class LabelController extends Controller
 
     public function destroy(Label $label): RedirectResponse
     {
+        if ($label->tasks()->exists()) {
+            flash()->error(__('label.has_tasks'));
+            return back()->withErrors([
+                'message' => __('label.has_tasks'),
+            ]);
+        }
+
         $label->delete();
         flash()->success(__('label.deleted'));
         return redirect(route('labels.index'));
