+++ app/Http/Controllers/TaskStatusController.php
@@ -13,7 +13,7 @@ class TaskStatusController extends Controller
 {
     public function __construct()
     {
-        $this->middleware('auth')->except('index');
+        $this->authorizeResource(TaskStatus::class);
     }
 
     public function index(): View
@@ -37,7 +37,7 @@ class TaskStatusController extends Controller
 
     public function edit(TaskStatus $taskStatus): View
     {
-        return view('task_status.edit', ['status' => $taskStatus,]);
+        return view('task_status.edit', ['status' => $taskStatus]);
     }
 
     public function update(UpdateTaskStatusRequest $request, TaskStatus $taskStatus): RedirectResponse
@@ -49,6 +49,13 @@ class TaskStatusController extends Controller
 
     public function destroy(TaskStatus $taskStatus): RedirectResponse
     {
+        if ($taskStatus->tasks()->exists()) {
+            flash()->error(__('task_status.has_tasks'));
+            return back()->withErrors([
+                'message' => __('task_status.has_tasks'),
+            ]);
+        }
+
         $taskStatus->delete();
         flash()->success(__('task_status.deleted'));
         return redirect(route('task_statuses.index'));
